added user authentication

+++ frontend/src/components/MeetingCard.jsx
@@ -1,4 +1,5 @@
 import { useMeetingContext } from '../hooks/useMeetingContext'
+import { useAuthContext } from '../hooks/useAuthContext'
 import { useState } from 'react'
 
 // icons
@@ -7,6 +8,7 @@ import cancelIcon from '../assets/cancel.svg'
 
 const MeetingCard = ({ meeting }) => {
     const { dispatch } = useMeetingContext()
+    const {user} = useAuthContext()
     const [expandMeeting, setExpandMeeting] = useState(false)
 
     const handleClick = () => {
@@ -14,8 +16,15 @@ const MeetingCard = ({ meeting }) => {
     }
 
     const delClick = async () => {
+        if (!user) {
+            return
+        }
+
         const response = await fetch('/api/meeting/' + meeting._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization':`Bearer ${user.token}`
+            }
         })
         const json = await response.json()
